new bug just dropped (numpy update?)

+++ src/napari_mm3/_subtract.py
@@ -103,7 +103,7 @@ def subtract_fluor(
                 empty_channel,
                 [
                     [
-                        np.int(0.5 * pad_row_length),
+                        np.int64(0.5 * pad_row_length),
                         pad_row_length - np.int(0.5 * pad_row_length),
                     ],
                     [
